\name{gfmodel11-package}
\alias{gfmodel11-package}
\alias{gfmodel11}
\docType{package}
\title{
\packageTitle{gfmodel11}
}
\description{
\packageDescription{gfmodel11}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gfmodel11}
\packageIndices{gfmodel11}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gfmodel11}

Maintainer: \packageMaintainer{gfmodel11}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
